backend/requirements.txt
# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# LangGraph for agent workflows (with compatible langsmith)
langgraph>=0.0.20
langsmith==0.1.125

# Tavily Search API
tavily-python>=0.2.0

# Gemini (Google Generative AI)
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Additional dependencies that might be needed
requests>=2.31.0

